pyproject.toml
[project]
name = "mediafusion"
version = "4.3.35"
description = "Media Fusion Add-On For Stremio & Kodi - A powerful streaming API with support for multiple providers and advanced scraping capabilities"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
authors = [
    { name = "Mohamed Zumair", email = "mhdzumair@gmail.com" },
    { name = "ragmehos" }
]

dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic[email]",
    "requests",
    "beautifulsoup4",
    "beanie==1.30.0",
    "motor",
    "pymongo",
    "dnspython",
    "jinja2",
    "apscheduler",
    "python-dateutil",
    "bencodepy",
    "pydantic-settings",
    "pycryptodome",
    "pillow",
    "httpx",
    "uvloop; sys_platform != \"win32\"",
    "thefuzz",
    "pikpakapi",
    "diskcache",
    "demagnetize",
    "redis[hiredis]",
    "dramatiq[redis,watch]",
    "gunicorn",
    "scrapy",
    "aioqbt",
    "aiowebdav",
    "m3u-ipytv",
    "python-multipart",
    "prometheus-client",
    "pyasynctracker",
    "scrapy-playwright",
    "cinemagoerng",
    "tqdm",
    "parsett",
    "tenacity",
    "ratelimit",
    "qrcode",
    "aioseedrcc",
    "pytz",
    "aiohttp-socks",
    "socksio",
    "dramatiq-abort",
    "tzdata",
    "ipython",
    "scrapy-fake-useragent>=1.4.4",
    "humanize>=4.12.1",
    "dateparser>=1.2.1",
    "feedparser>=6.0.11",
    "xmltodict>=0.14.2",
    "sqlmodel>=0.0.27",
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "typer>=0.20.1",
    "telethon>=1.37.0",
    "yt-dlp",
]

[dependency-groups]
dev = [
    "pysocks",
    "requests-cache",
    "kodistubs",
    "ruff>=0.8.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[tool.uv.sources]
pikpakapi = {git = "git+https://github.com/mhdzumair/PikPakAPI.git"}
aioqbt = {git = "git+https://github.com/mhdzumair/aioqbt.git"}
cinemagoerng = {git = "git+https://github.com/mhdzumair/cinemagoerng.git"}
demagnetize = {git = "git+https://github.com/mhdzumair/demagnetize.git"}

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "E712",   # comparison to True/False (needed for SQLAlchemy column comparisons)
]

[tool.ruff.lint.per-file-ignores]
# Migration scripts may have special requirements
"migrations/**/*.py" = ["E712"]
# Kodi plugin has different conventions
"kodi/**/lib/*.py" = ["F841"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
python_classes = ["Test*"]
addopts = ["-v", "--tb=short"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
